+++ checker/changes_test.go
@@ -47,3 +47,41 @@ func TestChanges_Count(t *testing.T) {
 	require.Equal(t, 0, lc[checker.WARN])
 	require.Equal(t, 1, lc[checker.ERR])
 }
+
+func TestIsEmpty_EmptyIncludeWarns(t *testing.T) {
+	bcErrors := checker.Changes{}
+	require.False(t, bcErrors.HasLevelOrHigher(checker.WARN))
+}
+
+func TestIsEmpty_EmptyExcludeWarns(t *testing.T) {
+	bcErrors := checker.Changes{}
+	require.False(t, bcErrors.HasLevelOrHigher(checker.ERR))
+}
+
+func TestIsEmpty_OneErrIncludeWarns(t *testing.T) {
+	bcErrors := checker.Changes{
+		checker.ApiChange{Level: checker.ERR},
+	}
+	require.True(t, bcErrors.HasLevelOrHigher(checker.WARN))
+}
+
+func TestIsEmpty_OneErrExcludeWarns(t *testing.T) {
+	bcErrors := checker.Changes{
+		checker.ApiChange{Level: checker.ERR},
+	}
+	require.True(t, bcErrors.HasLevelOrHigher(checker.ERR))
+}
+
+func TestIsEmpty_OneWarnIncludeWarns(t *testing.T) {
+	bcErrors := checker.Changes{
+		checker.ApiChange{Level: checker.WARN},
+	}
+	require.True(t, bcErrors.HasLevelOrHigher(checker.WARN))
+}
+
+func TestIsEmpty_OneWarnExcludeWarns(t *testing.T) {
+	bcErrors := checker.Changes{
+		checker.ApiChange{Level: checker.WARN},
+	}
+	require.False(t, bcErrors.HasLevelOrHigher(checker.ERR))
+}
